Add support for using Boost as the bignum library
